Move agencyTypeAndName to countyAgenciesSelectors where it belongs
- Note agencyTypeAndName is not a selector yet and should be refactored
to be one at a later time. Perhaps in story #150924573

[#146178793]
