+++ app/javascript/screenings/CrossReportShowView.jsx
@@ -5,26 +5,11 @@ import React from 'react'
 import ShowField from 'common/ShowField'
 import _ from 'lodash'
 import {dateFormatter} from 'utils/dateFormatter'
+import {agencyTypeAndName} from 'selectors/countyAgenciesSelectors'
 
 export default class CrossReportShowView extends React.Component {
   constructor() {
     super(...arguments)
-  }
-
-  agencyTypeAndName(agency_type, agency_code) {
-    const typeId = agency_type.replace(/ /gi, '_').toUpperCase()
-    const agencies = this.props.countyAgencies[typeId]
-    if (_.isEmpty(agencies)) {
-      return agency_type
-    } else {
-      const filteredAgencies = agencies.filter((countyAgency) => countyAgency.id === agency_code)
-      if (_.isEmpty(filteredAgencies)) {
-        return agency_type
-      } else {
-        const agencyName = filteredAgencies[0].name
-        return agencyName ? `${agency_type} - ${agencyName}` : agency_type
-      }
-    }
   }
 
   render() {
@@ -51,7 +36,7 @@ export default class CrossReportShowView extends React.Component {
                   {
                     crossReports.map(({agency_type, agency_name}, index) => (
                       <div key={index}>
-                        <li>{this.agencyTypeAndName(agency_type, agency_name)}</li>
+                        <li>{agencyTypeAndName(agency_type, agency_name, this.props.countyAgencies)}</li>
                         <ErrorMessages
                           errors={this.props.errors.getIn([agency_type, 'agency_name']) && this.props.errors.getIn([agency_type, 'agency_name']).toJS()}
                         />
