+++ app/javascript/selectors/countyAgenciesSelectors.js
@@ -1,4 +1,5 @@
 import {createSelector} from 'reselect'
+import _ from 'lodash'
 
 export const COMMUNITY_CARE_LICENSING = 'COMMUNITY_CARE_LICENSING'
 export const COUNTY_LICENSING = 'COUNTY_LICENSING'
@@ -23,3 +24,20 @@ export const getLicensingAgencies = createSelector(
   (state) => state.get('countyAgencies'),
   (countyAgencies) => countyAgencies.filter((countyAgency) => countyAgency.get('type') === COMMUNITY_CARE_LICENSING || countyAgency.get('type') === COUNTY_LICENSING)
 )
+
+// TODO: This should become a real selector like the above in the future say in #150924573
+export const agencyTypeAndName = (agency_type, agency_code, countyAgencies) => {
+  const typeId = agency_type.replace(/ /gi, '_').toUpperCase()
+  const agencies = countyAgencies[typeId]
+  if (_.isEmpty(agencies)) {
+    return agency_type
+  } else {
+    const filteredAgencies = agencies.filter((countyAgency) => countyAgency.id === agency_code)
+    if (_.isEmpty(filteredAgencies)) {
+      return agency_type
+    } else {
+      const agencyName = filteredAgencies[0].name
+      return agencyName ? `${agency_type} - ${agencyName}` : agency_type
+    }
+  }
+}
